GAME OF THE ERA
this is  a game that have html css and js code file that help to go the flash back of the child life that most famous and most loved game snake game
to make it more attractive their is a another css file have different name just edit in the html code css name and play and also mention your score.
my highest score is 49.
NOW ALL SET..